+++ lib/presentation/ui/widgets/upload_card_widget.dart
@@ -2,99 +2,10 @@ import 'package:flutter/material.dart';
 import 'package:iconsax/iconsax.dart';
 import 'package:share_flare/presentation/ui/utilities/assets_path.dart';
 import 'package:share_flare/presentation/ui/utilities/colors.dart';
+import 'package:share_flare/presentation/ui/widgets/comment_list_title_widges.dart';
+import 'package:share_flare/presentation/ui/widgets/comment_text_field.dart';
 
-class CommentListTile extends StatefulWidget {
-  @override
-  _CommentListTileState createState() => _CommentListTileState();
-}
 
-class _CommentListTileState extends State<CommentListTile> {
-  bool _heartInTap = false;
-
-  @override
-  Widget build(BuildContext context) {
-    return ListTile(
-      titleAlignment: ListTileTitleAlignment.top,
-      leading: Container(
-        height: 30,
-        width: 30,
-        decoration: BoxDecoration(
-          color: Colors.blue,
-          shape: BoxShape.circle,
-          image: DecorationImage(
-            image: AssetImage(SFAssetsPath.personPNG),
-            fit: BoxFit.cover,
-            onError: (Object, StackTrace) {
-              const Icon(Icons.image);
-            },
-          ),
-        ),
-      ),
-      title: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Row(
-            children: [
-              Text(
-                'mferdous12',
-                style: TextStyle(
-                  color: Color(0xFF1D2939),
-                  fontSize: 12,
-                  fontWeight: FontWeight.w500,
-                  height: 0.14,
-                ),
-              ),
-              SizedBox(width: 12.0,),
-              SizedBox(
-                width: 80,
-                child: Text(
-                  '2 hours ago',
-                  style: TextStyle(
-                    color: Color(0xFF98A2B3),
-                    fontSize: 12,
-                    fontWeight: FontWeight.w500,
-                    height: 0.14,
-                  ),
-                ),
-              )
-            ],
-          ),
-          SizedBox(height: 12.0,),
-          Text(
-            'Nice picture you have captured 🔥 ',
-            style: TextStyle(
-              color: Color(0xFF1D2939),
-              fontSize: 14,
-              fontWeight: FontWeight.w400,
-              height: 1,
-            ),
-          ),
-          SizedBox(height: 16.0,),
-          Text(
-            'Replay',
-            style: TextStyle(
-              color: Color(0xFF475467),
-              fontSize: 12,
-              fontFamily: 'Inter',
-              fontWeight: FontWeight.w400,
-              height: 0.12,
-            ),
-          ),
-        ],
-      ),
-      trailing: GestureDetector(
-        onTap: () {
-          setState(() {
-            _heartInTap = !_heartInTap;
-          });
-        },
-        child: _heartInTap
-            ? const Icon(Iconsax.heart5, size: 20, color: SFColors.heartRedColor)
-            : const Icon(Iconsax.heart, size: 20),
-      ),
-    );
-  }
-}
 
 class UploadCardWidget extends StatefulWidget {
   const UploadCardWidget({Key? key}) : super(key: key);
@@ -314,7 +225,7 @@ class _UploadCardWidgetState extends State<UploadCardWidget> {
                         padding: EdgeInsets.zero,
                         itemCount: 10,
                         itemBuilder: (context, index) {
-                          return CommentListTile();
+                          return CommentListTileWidget();
                         }),
                   ),
                   Divider(color: Colors.grey.withOpacity(0.2),),
@@ -355,61 +266,5 @@ class _UploadCardWidgetState extends State<UploadCardWidget> {
   }
 }
 
-class CommentTextFieldWidget extends StatefulWidget {
-   CommentTextFieldWidget({Key? key}) : super(key: key);
 
-  @override
-  State<CommentTextFieldWidget> createState() => _CommentTextFieldWidgetState();
-}
-
-class _CommentTextFieldWidgetState extends State<CommentTextFieldWidget> {
-  TextEditingController _commentTEController = TextEditingController();
-
-  bool _changeText = false;
-
-  @override
-  Widget build(BuildContext context) {
-    return TextFormField(
-      controller: _commentTEController,
-      onChanged: (String value) {
-        if (value.isNotEmpty) {
-          _changeText = true;
-        } else {
-          _changeText = false;
-        }
-        setState(() {});
-      },
-      decoration: InputDecoration(
-        contentPadding: const EdgeInsets.symmetric(
-            vertical: 4, horizontal: 16.0),
-        hintText: 'Add comments',
-        hintStyle: const TextStyle(
-          color: Color(0xFF98A2B3),
-          fontSize: 14,
-          fontWeight: FontWeight.w500,
-          height: 0.10,
-        ),
-        border: const OutlineInputBorder(
-           borderSide:BorderSide.none,
-        ),
-        suffixIcon: GestureDetector(
-          onTap: () {
-            _commentTEController.clear();
-            _changeText = false;
-            setState(() {});
-          },
-          child: SizedBox(
-            width: 15.0,
-            height: 14.0,
-            child: Center(
-              child: _changeText == true
-                  ? Icon(Iconsax.send_1)
-                  : null,
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-}
